msm: dcvs: update dcvs if the governor limits change
We had seen issues where dcvs goes out of sync with the actual
freq the cpu is running at. The root cause was if the userspace
changes the limits on the governor, the governor ends up changing
the frequency without notifying dcvs.

Provide an api for the governor to call dcvs when a frequency change
happens.

Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
(cherry picked from commit 68c970e38fcb700d851301dbacc56da9387294c6)

Change-Id: Ib55f1f2061c385bb68281263a2bfcc9c20e4cf39
Signed-off-by: Ram Kumar Chakravarthy Chebathini <rcheba@codeaurora.org>